RT #3594: Added @url: articles_path@ option to @form_with@.
The articles_path helper tells Rails to point the form to the URI Pattern associated with the articles prefix; and the form will (by default) send a POST request to that route. This is associated with the create action of the current controller, the ArticlesController.

With the form and its associated route defined, you will be able to fill in the form and then click the submit button to begin the process of creating a new article.