RT #3594: Refactoring: added partial for edit and new forms - these are identical so they may use single partial for displaying form data.
Everything except for the form_with declaration remained the same. The reason we can use this shorter, simpler form_with declaration to stand in for either of the other forms is that @article is a resource corresponding to a full set of RESTful routes, and Rails is able to infer which URI and method to use. For more information about this use of form_with, see "Resource-oriented style":http://api.rubyonrails.org/v5.2.3/classes/ActionView/Helpers/FormHelper.html#method-i-form_with-label-Resource-oriented+style.