RT #3594: Added @ArticlesController#destroy@ action.
Here we're using link_to in a different way. We pass the named route as the second argument, and then the options as another argument. The method: :delete and data: { confirm: 'Are you sure?' } options are used as HTML5 attributes so that when the link is clicked, Rails will first show a confirm dialog to the user, and then submit the link with method delete. This is done via the JavaScript file rails-ujs which is automatically included in your application's layout (app/views/layouts/application.html.erb) when you generated the application. Without this file, the confirmation dialog box won't appear.